Support `credential-process` for use by awscli configured profiles
Support a `credential-process` command for supplying credentials to `awscli:credential_process`.

awscli allows for supplying credentials to a configured profile using a command provided by `credential_process` in `~/.aws/config`. The command is run by the tool, and the credentials are provided in a json format to be read by the tooling.

The usage in `~/.aws/config` can be seen below, with the authentication and account details prompted for, and supplied interactively.

```
[profile master]
output = json
credential_process = bmx credential-process --org acme --user jonathan.beverly
```

 The prompts require writing to `Stdin` in the credential process case, but this has been expanded to all prompts. Although this works in my local tests, I mark this as a non-production viable option. The reason it is likely working locally is that `Stdin` is just echoing back the result into the terminal, and this implementation will encounter compatibility issues when attempting to roll out into multiple environments (docker as an example).

For now I'm fine with the stdin implementation, as it allows me to progress onto the next requirements for getting bmx compatible with a MacOS + okta push environment.