Limit custom tty usage to credential-process command
Limit the exposure of the new tty component to only the credential-process command.

To do this, it requires adding in the concept of `state` to the DefaultConsoleReader. For this state to be reflected in all usages, a single console reader is now created at the root, and passed down to all components that require it. This has had some consequences with refactoring the location of object initialization. I'm comfortable with moving forward with these changes though.

This `/dev/tty` has only been confirmed working on my local environment, and will need to be confirmed for other scenarios such as git commit hooks, docker and different terminals.

As an aside, the semantics of the reader is a little off, as it is now able to read & write. I haven't renamed the file yet, as this is still in an experimental state, and the writer component may be split to its own struct.