Create the session storage directory if not exists
Create the session storage directory if it does not exist, allowing use of session in subsequent runs.

This resolves an issue when running in the context of a git commit hook. Specifically, when calling a command within the hook that handles its own credential assumption, it will not have a valid device (inappropriate ioctl) to work with. The calls from the commit hook script itself can be interactive and handle authentication.

By ensuring that the directory is created after authentication, this helps reduce the likelihood of encountering scenarios where a pipe is unavailable.