License adjustments in the source code & minor fixes
Fill in the LICENSE with missing corporation & year inputs pulled from file headers.

The apache license version 2.0 documentation covers how to apply the license to your work: https://www.apache.org/licenses/LICENSE-2.0#apply. Based on this process, I have updated the LICENSE itself with the year and corporation outputs.

A note on file headers for each of the sources. This does not appear to be necessary for Apache2 unless you are intending to distribute files individually included in this project (in some manner). The exact requirement can be read in the policy https://www.apache.org/legal/src-headers.html (or FAQ: https://infra.apache.org/apply-license.html#copy-per-file)

Although these headers are likely not needed, they cannot be removed unless by an member of D2L Corporation as per the license requirements of Apache2.