fix(ineffassign): Ineffectual assignments of err responses in okta provider (#5)
Resolve a series of ineffectual error assignments in the Okta provider.

This was caught by goreportcard-cli, when run against the project. The style of err handling is mirrored by similar if statements in the area.

Catching these errors will need to be bundled into CI at a later date to prevent them from occurring in the future.