feat(version): Stamp compiled binary with version and commit (#7)
Stamp the bazel compiled binaries (missing goreleaser binaries) with the version and commit sha.

For debugging purposes it will be good to have a commit sha associated with the binary, so it can referenced back if its an out of date version of `bmx`. For cases that are installed with `go get`, the version will not be stamped, which will yield the `bmx/nostamp` output.