feat(cmd-login): Login command for creating Okta sessions (#9)
Introduce `login` command for creating a new session with Okta.

Sessions are created in the process of executing other commands (e.g. `print`, `credential-process`), while we have no explicit command for creating a new session. This enables an option for creating a session without needing to ignore the response.

This is related to an issue when developing when working in a non-interactive environment. In those cases, we need to login ahead of time to avoid any prompts. At the moment that can be done using `bmx print` and discarding the output (/dev/null).

The command inputs of `login` mirror that of `print` to match the UserInfo API. This can be refactored in future work to reduce the inputs.