feat(cmd-config): Emit default config for user directory (#10)
Emit an ini configuration file for mapping to a Okta organization and user.

This is intended to assist with automated configured of user environments by allowing for bmx to create a simple configuration file for a user. This doesn't do any mappings for other config options like account, which would need to be set by the user. When there is only a single AWS okta app, then bmx should default to selecting that option.

This allows for cases where the machine username matches the okta username, you can run this:

```
bmx config --org 'ACME' --user `id -n -u` > ~/.bmx/config
```

To fill in the details of the organization and user.