fix(console): Print newline after reading password (#14)
Print newline after reading password to fix line breaks for prompts.

Any text that follows the password prompt will land on the same line as the password prompt. Visually this can be jarring as it doesn't make it immediately obvious what the next prompt is asking (as its in an odd spot). This resolves this by adding a new line after the password is retrieved.

See below for an example of odd placement.

```
Okta Password:  Available MFA: 
```