docs(acknowledge): Missing acknowledgement of D2L (#17)
Add in a clear acknowledgement of the work done by D2L from the original project (Brightspace/bmx).

The project is missing a clear indication that D2L was the sole developer of bmx before modifications were made by members of Arctic Wolf (@rtkjbeverly). The project wasn't forked when it was originally experimented with during a hackathon, and until now it hasn't been noticed that the project was missing acknowledgements of this.

This adds in acknowledgements of this, but isn't complete.