feat(assume): Support an assume role action from print (#19)
Support role assumption from the Okta SAML role from the print command.

When working with `bmx print` the SAML role is not always the intended role for credentials. This means that either AWS profiles or a `aws sts assume-role` is necessary to get the intended role STS credentials. This change allows for print to perform an AssumeRole action from the SAML role, for returning the STS credentials. 

Users can now configure the configuration property `AssumeRole` in a project context. This allows a `.bmx` file to exist in a directory with the contents `AssumeRole = ...`. When print is executed, the role specified in the file will be used when assuming the role.

Additionally this can be set at the user level in cases where a user may only have access to a single workable role.

Co-authored-by: Chris Stavropoulos <chris.stavropoulos@arcticwolf.com>