refactor(saml-roles): Split SAML role selection from GetCredentials (#23)
Split the selection of a SAML role from GetCredentials to better use `Option`.

This splits the selection of a SAML role from GetCredentials, moving the selection process out and passing in a resolved role. This refactoring was done as Option is needed in `pickRole` for the implementation of AppleScript in the MacOS case. While looking into the steps necessary for this, it made sense to move the role selection out of GetCredentials, and require a resolved SAML role to be passed in.

This does introduce issues where the concept of `AWS` is leaking from the serviceProvider abstraction. I'm comfortable with this for now to support alternative options of selection like AppleScript for MacOS. Fixes for this abstraction will come in later PRs.