fix(totp): TOTP code cannot have trailing whitespace (#25)
The TOTP code sent to Okta cannot have trailing whitespace or it will fail to verify.

ReadLine does not trim trailing whitespace from inputs. For the console case this doesn't happen often, but for other input devices there may be trailing whitespace. ReadLine itself shouldn't trim whitespace as that may be intended for the input.