refactor(console): Refactor console for multiple input modes (#26)
Refactor console interface to better handle different input models.

The DefaultConsoleReader was built around working with terminal interfaces like zsh/bash/powershell. When support for `credential-process` was added, it was built with a narrow focus on usage from a console. This isn't always the case, and we need to be able to be flexible with where input is coming from, as `credential_process` isn't always invoked in the context of an interactive shell.

This lays the ground-work for supporting AppleScript in MacOS environments, which is the primary driver of this work.