feat(console-applescript): Support for AppleScript input while on MacOS (#27)
Support AppleScript as an input option while working from MacOS.

This adds support for AppleScript in MacOS for dealing with cases where `bmx` may be running from a non-interactive context. When this happens, AppleScript will be selected instead of console input.

The configuration enum `input` has been made available to change the behaviour of input device selection. This is intended to address scenarios where the determination of an available `tty` is incorrect, and the interface option needs to be overwritten.

The available options are [`console`, `applescript`, `always_applescript`], with the options described as such:

- `console` tries to use the console first, then falls back on other input methods.
- `applescript` tries to use the console first, and uses AppleScript first when in limited.
- `always_applescript` tries to use AppleScript for input always.

The only limited context in the code at the moment is the command `credential-process`.