doc(input): Configuration enum to control input behaviour (#28)
Document the configuration enum `input` that controls input behaviour.

To accommodate some edge cases of `bmx` running in cases without access to a terminal, there is an input config that assists in controlling how input is sourced. This documents the options, and how they work in the system as a whole.