build(bazel): nogo to run static analysis on build (#29)
Enable nogo in the bazel build process for performing static analysis on the code.

This enables static analysis to be performed in the `bazel build //` process for bmx. This will ensure `go vet` is run, and that any issues with the code is caught early. Duplicate entries for the static analysis tools have been removed from the list. When github actions are switched from using golang to bazel, this will simplify the static analysis and build steps.

Additionally the file `.bazelversion` has been added to pin the bazel verison for bazelisk.