build(devcontainers): Support devcontainer environments (#32)
Enables a bazel devcontainer environment that can be used for isolated development.

This enables development in a docker container, which allows for isolation from the current `bmx` installation on a developer machine. This can be useful when attempting to avoid sessions or config clashes with already configured environments.

Bazel buildtools like buildifier and buildozer have been installed on the machine. The list of tools is in no way complete.