legal(license): Add AWN to license notice (#30)
Add Arctic Wolf Networks to the repository license to reflect code changes by AWN.

The project license remains the same, but the attribution for work done by Arctic Wolf is added to the root license. 

The file headers in all cases have not been updated as part of this, as that should be done separately. My understanding is that the license headers are only a recommendation, and not a requirement of license. However as this is in a copyright realm, I'm erring on the side of absurd levels of caution.