build(github-actions): Use bazel for builds (#34)
Replace the golang-based build & test process with bazel.

This switches from configuring golang on the build machine for build & test, to using the bazel-driven process. The local development experience should now more closely match with that of the continuous integration build & test. This has the added benefit of validating that the bazel build process is successfully building the tool.