build(vendor): Vendor the dependencies of tool (#35)
Vendor the depedencies of bmx, using `bazel gazelle` to manage them in the build process.

This vendors the golang dependencies into the repository, using bazel to manage them in the build process and `go mod` for the dependency upgrade process. The documentation has been updated to reflect this change.

As a consequence of this, the build times for CI has increased but still remains within the expected 5 minute limit.