build(goreleaser): Remove the goreleaser validation (#37)
GoReleaser is being removed in favour of using the Bazel artifacts.

This removes the goreleaser validation workflow in preparation for the removal of GoReleaser for building the deployment artifacts. Using GoReleaser for the compiled artifacts means that the local and CI artifacts are built with a different process than the deployment artifacts. This hasn't yielded any problems so far, but adds additional tools that are not needed.