fix(applescript): AppleScript prompts can raise concerns (#38)
AppleScript prompts do not clearly indicate that they are from BMX, raising suspicions.

This add BMX to the AppleScript dialog box titles to ensure that users are aware of the context behind the prompt.

This has shown up when individuals open IntelliJ, and are suddenly prompted for their Okta password with no context behind what script/utility is doing the prompt. Individuals with context of the prompt, may understand that it is done by BMX, but this conditions users to input credentials into unknown/unclear prompts. We'd like to avoid that as much as possible.