remove(build): Use bazel for release bundles (#39)
Use the bazel builds for release bundles instead of relying on goreleaser to generate the files.

This deprecates the old goreleaser way of creating release bundles in favour of using the compiled bazel releases. This does mean that checksums are no longer included with releases, and the goreleaser conventions are no longer applied to github releases.