feat(logout): Clear stored sessions (#41)
Introduce a `logout` command for clearing the sessions file.

This introduces a `logout` command to have a managed way of clearing out the `sessions` file. There are some edge cases, where an invalid token (or soon to be expired token) is used to communicate with Okta. The solution at present is to clear or delete the sessions file (`~/.bmx/sessions`).

The `logout` command adds a more managed way of doing this, till the exact cause of this issue can be resolved.