fix(credential-processs): Remove tty available check
**Description:** Rather than check for the ability to create a tty
connection and failing if we can't, continue as if we have one. In the
case where we have a valid session, we don't need the tty, so the
credential-process command will continue to work. If we need to use the
tty (because no valid session), we already are raising an error, so let
it fail then.

**Testing:** `bazel test //...` and manually testing locally