fix(applescript): Support empty string default (#52)
Support empty string defaults for the AppleScript dialog boxes.

This resolves an issue where if the Answer field was set as the empty string, it would no longer show the textbox. This allows a textbox-always handling for empty string answers, to ensure the textbox exists. As we don't leverage the dialog for confirmation windows, this is sufficient to handle out input case.

Confirmed working by testing locally with the AppleScript base cases (push/totp, no username, etc).

Additionally the location of the config file is now using the projectFileName constant.