fix(build): Local source from `go get` using wrong package (#54)
When installing bmx using go get, it is using the github.com version of mack instead of vendored.

This moves the vendored version of mack into the source, rather than treat it as a vendored dependency. As this way it permits easier modification of the library, and clarifies that it is not expected to match exactly with the source.